+++ tests/TestCase.php
@@ -4,5 +4,4 @@ use PHPUnit\Framework\TestCase as BaseTestCase;
 
 class TestCase extends BaseTestCase
 {
-
 }